refactor: improve ci speed by parallelization
3 test suites
